[#1403] moved method to another file
