+++ data/src/main/java/org/akvo/flow/data/util/FileHelper.java
@@ -24,28 +24,14 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.text.TextUtils;
 import android.util.Base64;
-
-import java.io.BufferedInputStream;
-import java.io.Closeable;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-import java.util.zip.Adler32;
-import java.util.zip.CheckedOutputStream;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipInputStream;
-import java.util.zip.ZipOutputStream;
-
-import javax.inject.Inject;
-
 import okhttp3.ResponseBody;
 import timber.log.Timber;
+
+import javax.inject.Inject;
+import java.io.*;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+import java.util.zip.*;
 
 public class FileHelper {
 
@@ -204,6 +190,19 @@ public class FileHelper {
         InputStream inputStream = responseBody.byteStream();
         extractZipContent(inputStream, targetFolder);
         close(inputStream);
+    }
+
+    public boolean validFile(File file) {
+        return file.exists() && validZipFile(file);
+    }
+
+    private boolean validZipFile(File file) {
+        try {
+            return new ZipFile(file).size() > 0;
+        } catch (IOException e) {
+            Timber.e(e);
+            return false;
+        }
     }
 
     private void copyStream(InputStream inputStream, File destinationFile) {
